requirements/base.txt
Django==2.2.1
django-cors-headers==2.5.3
djangorestframework==3.9.3
# Mysql
mysqlclient==1.4.2.post1